OMAP3: SGX-KM: UDD: New debug messaging scheme in 3PDD
A new level of messaging called INFO has been added
in the 3rd Party Display Driver to display basic
information about the creation of the swap chain
and the number of buffers available.

This patch also allows to enable in the configuration
menu of the kernel, debug messages from the 3rd Party
Display Driver without the need of building the SGX
driver in debug mode.

Change-Id: I3029a360064882cd96b08af137afb4bb291655cf
Signed-off-by: Rodrigo Obregon <robregon@ti.com>