OMAP3:PM:Re-order MPU and DSP freq scaling
Now, while scaling from VDD1 OPPx-->OPP5, we need to make sure
that DSP freq is changed first and then the ARM. Else, there can
be a small window were ARM will be running at 1.2G with DSP greater
than 65MHz. Which is not a stable combination as per h/w validation.
So, this patch makes sure to change the DSP freq first (if the target OPP is
OPP5) and then the ARM.
And for other OPP transition, change the ARM freq and then the DSP.

Change-Id: I196f05ce0e2bdc87291d3ad44b955de26ec1efa0
Signed-off-by: Girish S G <girishsg@ti.com>